+++ src/confusionMatrix_.cpp
@@ -117,3 +117,27 @@ double mcc_(NumericVector actual, NumericVector predicted, double cutoff){
 }
 
 
+// [[Rcpp::export]]
+double kappa_(NumericVector actual, NumericVector predicted, double cutoff){
+
+  // True Negatives d
+  double TN = sum(predicted < cutoff & actual == 0);
+  // False Negatives - c
+  double FN = sum(predicted < cutoff & actual == 1);
+  // False positives - b
+  double FP = sum(predicted >= cutoff & actual == 0);
+  // True positives - a
+  double TP = sum(predicted >= cutoff & actual == 1);
+
+  double N = (TP + FP + FN + TN);
+
+  double po = (TP + TN)/N;
+  double margin_a = ((TP + FP)*(TP + FN))/N;
+  double margin_b = ((FN + TN)*(FP + TN))/N;
+
+  double pe = (margin_a + margin_b)/N;
+
+  double kappa = (po - pe)/(1 - pe);
+  return kappa;
+}
+