+++ NEWS.md
@@ -1,5 +1,6 @@
 # ModelMetrics 1.2.0
 
+* added kappa statistic
 * added s3 methods for `glm`, `lm`, `randomForest`, `merMod`, and `glmerMod`
 * sped up `auc` with `data.table::frankv`
 * added `gini`